rename units.py to unit.py #246
