+++ test/unit/repositories/test_unit.py
@@ -13,7 +13,7 @@ from inventory_management_system_api.core.exceptions import (
     DuplicateRecordError,
     InvalidObjectIdError,
 )
-from inventory_management_system_api.models.units import UnitIn, UnitOut
+from inventory_management_system_api.models.unit import UnitIn, UnitOut
 
 
 def test_create(test_helpers, database_mock, unit_repository):
